Feature/Scrum 105: Implementar configuracion de tests unitarios (#7)
* Agregar test service a docker-compose y crear configuración pytest

* Primer intento de testing unitario, falta configurar docherfile.dev

* Cambios para simplificar setup de tests

* Agregar la funcionalidad para CRUD de Usuario (#4)

* Commit inicial

* Agregar CRUD de clientes, asociar clientes a usuarios

* Agregar validación requerida al campo "Tipo de Cliente" en el formulario de CRUD de cliente

* Inicio de actualizacion de html/css + documentacion de codigo

TODO:
Corregir build de css de tailwind, para que encuentre las rutas de templates de html en apps/
Actualizar templates asociar_cliente_a_usuario.html, cliente_list.html, rol_list.html a usar formato de header/footer/base.html definido para el proyecto en app/templates

* Refactorizar templates HTML para extender de base.html, agregar rutas al input.css

Los templates HTML ahora extienden del base.html para utilizar el mismo header y footer. Se agregaron más rutas al input.css para generar correctamente el output.css

* Arreglar autoformateo de templates

---------


* Feature/Scrum 108: Configurar directorios y estructura de proyecto Django (#6)

* Agregar carpeta IA_Chats con link_chats.txt

* Agregar apps y urls de ejemplo para nuestros modulos

* Primer intento de testing unitario, falta configurar docherfile.dev

* Cambios para simplificar setup de tests

* Tests unitarios configurados, falta confirmar con una primera prueba.

* Saco el create shell  (sh -c " ") de docker compose dev

* Agregar testUsuarioModel temporal para comprobar funcionamiento de tests

* Feature: Implementar docs automaticos con Sphinx (#8)

* Indice de sphinx funcionando para todas las apps de Django
* Github Actions para construir la documentacion, y desplegar a github pages en push a develop


Commits:
* Update pyproject.toml

* Update pyproject.toml

* Update pyproject.toml

* Update pyproject.toml

* cambios para evitar error: entrypoint  permission denied

* hice gitflow init y cree la rama del scrum59, creo que eso causó la necesidad de este commit

* antes de continuar, voy a crear los directorios para cada módulo contemplado
y los archivos __init__.py para que python los reconozca como paquetes

* creación de la rama SCRUM80 para crear los directorios de  los módulos contemplados

* feat: Estructura para la documentacion automatica

* fix: Corregido el path de Sphinx en conf.py

* feat: Agregar configuracion de pylance en vscode

* feat: Agregar configuracion de pylance en vscode

* feat: Configurado Pylance en VS Code para leer variables de entorno

* fix: Solucionado el error de ALLOWED_HOSTS en la configuración de Sphinx

* fix: Corregido el error de importación de módulos en la configuración de Sphinx

* fix: Corregido el error de importación de módulos en la configuración de Sphinx

* fix: Corregida la ruta en conf.py para la importación de módulos

* fix: Corregido el error de importación de módulos en la configuración de Sphinx

* Agregar carpeta IA_Chats con link_chats.txt

* Agregar apps y urls de ejemplo para nuestros modulos

* Eliminar node_modules del PR

* Borrar mas archivos extra

* Cambiar github actions CI/CD

---------


* Agregar test service a docker-compose y crear configuración pytest

* Primer intento de testing unitario, falta configurar docherfile.dev

* Cambios para simplificar setup de tests

* Agregar la funcionalidad para CRUD de Usuario (#4)

* Commit inicial

* Agregar CRUD de clientes, asociar clientes a usuarios

* Agregar validación requerida al campo "Tipo de Cliente" en el formulario de CRUD de cliente

* Inicio de actualizacion de html/css + documentacion de codigo

TODO:
Corregir build de css de tailwind, para que encuentre las rutas de templates de html en apps/
Actualizar templates asociar_cliente_a_usuario.html, cliente_list.html, rol_list.html a usar formato de header/footer/base.html definido para el proyecto en app/templates

* Refactorizar templates HTML para extender de base.html, agregar rutas al input.css

Los templates HTML ahora extienden del base.html para utilizar el mismo header y footer. Se agregaron más rutas al input.css para generar correctamente el output.css

* Arreglar autoformateo de templates

---------


* Feature/Scrum 108: Configurar directorios y estructura de proyecto Django (#6)

* Agregar carpeta IA_Chats con link_chats.txt

* Agregar apps y urls de ejemplo para nuestros modulos

* Primer intento de testing unitario, falta configurar docherfile.dev

* Cambios para simplificar setup de tests

* Tests unitarios configurados, falta confirmar con una primera prueba.

* Saco el create shell  (sh -c " ") de docker compose dev

* Agregar testUsuarioModel temporal para comprobar funcionamiento de tests

* Configurar Tests en github actions  tests.yml

* Probar arreglar CI

* Arreglar ruta archivo compose

---------
