--- .env.layout
@@ -1,2 +1,0 @@
-[Editors]
-Order=
