+++ Makefile
@@ -14,3 +14,7 @@ migrate:
 
 logs:
 	docker compose -f docker-compose.dev.yml logs -f
+
+# UV
+test: # Correr testeos unitarios de pytest en local (Se debe usar sqlite, TODO: configurar mejor)
+	uv run --env-file ..\.env.dev pytest
