+++ app/global_exchange_django/asgi.py
@@ -1,5 +1,4 @@
-"""
-ASGI config for global_exchange_django project.
+"""ASGI config for global_exchange_django project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
 
@@ -11,6 +10,6 @@ import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'global_exchange_django.settings')
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "global_exchange_django.settings")
 
 application = get_asgi_application()
