+++ app/global_exchange_django/wsgi.py
@@ -1,5 +1,4 @@
-"""
-WSGI config for global_exchange_django project.
+"""WSGI config for global_exchange_django project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +10,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'global_exchange_django.settings')
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "global_exchange_django.settings")
 
 application = get_wsgi_application()
