+++ app/pytest.ini
@@ -1,0 +1,4 @@
+# pytest.ini
+[pytest]
+DJANGO_SETTINGS_MODULE = global_exchange_django.settings
+python_files = tests.py test_*.py *_tests.py
