+++ app/static/css/tailwind.config.js
@@ -1,13 +1,13 @@
 module.exports = {
-    content: [
-        "../../templates/**/*.html",
-        "../../**/templates/**/*.html",
-        "../../apps/panel_admin/templates/**/*.html",
-        "../../**/apps/panel_admin/templates/**/*.html",
+  content: [
+    "../../templates/**/*.html",
+    "../../**/templates/**/*.html",
+    "../../apps/panel_admin/templates/**/*.html",
+    "../../**/apps/panel_admin/templates/**/*.html",
 
-    ],
-    theme: {
-        extend: {},
-    },
-    plugins: [],
+  ],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
 }
