+++ app/templates/base.html
@@ -2,26 +2,29 @@
 {% load static %}
 <!DOCTYPE html>
 <html lang="es">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width,initial-scale=1">
-    <title>
-      {% block title %}
-        Global Exchange
-      {% endblock title %}
-    </title>
-    <!-- global css - Tailwind/daisy output -->
-    <link rel="stylesheet" href="{% static 'css/output.css' %}">
-    {% block extra_head %}{% endblock %}
-  </head>
-  <body class="{% block body_class %}{% endblock %}">
-    {% block header %}
-      {% include "includes/header.html" %}
-    {% endblock %}
-    <main>
-      {% block content %}{% endblock %}
-    </main>
-    {% include "includes/footer.html" %}
-    {% block extra_scripts %}{% endblock %}
-  </body>
+<head>
+  <meta charset="utf-8" />
+  <meta name="viewport" content="width=device-width,initial-scale=1" />
+  <title>{% block title %}Mi Aplicación{% endblock %}</title>
+
+  <!-- CSS global (tailwind + daisyUI build) -->
+  <link rel="stylesheet" href="{% static 'css/output.css' %}">
+
+  {% block extra_head %}{% endblock %}
+</head>
+<body class="{% block body_class %}{% endblock %}">
+  {# Por defecto incluimos el header global, pero cada plantilla se puede
+     sobrescribir el bloque 'header' para usar otro header. #}
+  {% block header %}
+    {% include "includes/header.html" %}
+  {% endblock %}
+
+  <main>
+    {% block content %}{% endblock %}
+  </main>
+
+  {% include "includes/footer.html" %}
+
+  {% block extra_scripts %}{% endblock %}
+</body>
 </html>
