+++ docker-compose.dev.yml
@@ -10,12 +10,24 @@ services:
 
   web:
     build: ./app
-    command: uv run manage.py runserver 0.0.0.0:8000
+    command: uv run python manage.py runserver 0.0.0.0:8000
     working_dir: /app
     # volumes: TODO: montar entorno desarrollador, watch
     #   - ./app/:/app/
     ports:
       - 8000:8000
+    env_file:
+      - ./.env.dev
+    depends_on:
+      - db
+
+# Nuevo servicios para tests
+  test:
+    build: ./app
+    command: uv run python -m pytest
+    working_dir: /app
+    # volumes:
+    #   - ./app/:/app/
     env_file:
       - ./.env.dev
     depends_on:
